+++ src/app/weather/weather-region/weather-region.component.scss
@@ -1,4 +1,5 @@
 .btn-hover:hover {
 	background-color: #007bff;
 	color:white;
+	box-shadow:none;
 }
