+++ src/app/weather/weather-region/weather-region.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { WeatherRegions } from '../shared/classes/weather';
-import { Regions } from '../shared/classes/constants';
+import { WeatherRegions } from '../../shared/classes/weather';
+import { Regions } from '../../shared/classes/constants';
 
 @Component({
   selector: 'app-weather-region',
