+++ src/app/weather/weather-simulation/weather-simulation.component.ts
@@ -1,4 +1,3 @@
-
 import { Component } from '@angular/core';
 
 @Component({
