+++ src/app/weather/weather-sources/weather-sources.component.ts
@@ -1,4 +1,3 @@
-
 import { Component } from '@angular/core';
 
 @Component({
