+++ src/app/weather/weather_main/weather.component.scss
